mcp/quick-start/mcp-client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.58.2",
    "mcp>=1.12.0",
    "python-dotenv>=1.1.1",
]
